refactor(projects): extract row primitives to equipment-rows.tsx (pure move)
Phase 3a of cross-type unification. Moves the row/group rendering components
out of the 2767-line equipment-tab.tsx into a dedicated equipment-rows.tsx:
- types: LineItemData, GroupData, CategoryData, OverbookedInfo
- COL_COUNT, isRealKitChild, isHiddenFromList, isMergeTombstone
- OverbookedBadge, SortableGroupRow, SortableCategoryRow, SortableLineItemRow

Pure move, zero behaviour change — verified: the extracted 571-line block
diffs against the original by ONLY the 10 added `export` keywords, nothing
else. Trimmed the now-unused imports from equipment-tab (useSortable, CSS,
DropdownMenu*, Badge, Tooltip*, several icons). equipment-tab.tsx drops from
2767 to ~2190 lines.

Build + tsc + eslint clean (2 pre-existing warnings carried over verbatim).

Next commit: introduce the 3-axis RowDescriptor + rename to LineItemRow/GroupRow.
